# public/api-docs/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/cidadao": {
      "get": {
        "description": "Retorna uma lista paginada de cidadãos com opções de filtro.",
        "operationId": "Cidadao_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Termo de busca (busca por nome, CPF ou NIS)",
            "schema": {
              "example": "Maria",
              "type": "string"
            }
          },
          {
            "name": "bairro",
            "required": false,
            "in": "query",
            "description": "Filtrar por bairro",
            "schema": {
              "example": "Centro",
              "type": "string"
            }
          },
          {
            "name": "ativo",
            "required": false,
            "in": "query",
            "description": "Filtrar por status (true para ativos, false para inativos)",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Número de itens por página (padrão: 10, máximo: 100)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número da página (padrão: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de cidadãos retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CidadaoPaginatedResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar cidadãos",
        "tags": [
          "Cidadão"
        ]
      },
      "post": {
        "description": "Cadastra um novo cidadão no sistema.",
        "operationId": "Cidadao_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCidadaoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cidadão criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CidadaoResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "Cidadão criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CidadaoResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos fornecidos",
            "schema": {
              "example": {
                "statusCode": 400,
                "message": "CPF inválido",
                "error": "Bad Request"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflito - CPF ou NIS já cadastrado",
            "schema": {
              "example": {
                "statusCode": 409,
                "message": "Já existe um cidadão cadastrado com este CPF",
                "error": "Conflict"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar cidadão",
        "tags": [
          "Cidadão"
        ]
      }
    },
    "/api/v1/cidadao/{id}": {
      "get": {
        "description": "Retorna os detalhes completos de um cidadão pelo seu ID.",
        "operationId": "Cidadao_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do cidadão",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cidadão encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CidadaoResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "ID inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Cidadão não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter detalhes de um cidadão",
        "tags": [
          "Cidadão"
        ]
      },
      "put": {
        "description": "Atualiza os dados de um cidadão existente.",
        "operationId": "Cidadao_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do cidadão a ser atualizado",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCidadaoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cidadão atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CidadaoResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Cidadão não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "CPF ou NIS já em uso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar cidadão existente",
        "tags": [
          "Cidadão"
        ]
      }
    },
    "/api/v1/cidadao/cpf/{cpf}": {
      "get": {
        "description": "Busca um cidadão pelo número do CPF (com ou sem formatação).",
        "operationId": "Cidadao_findByCpf",
        "parameters": [
          {
            "name": "cpf",
            "required": true,
            "in": "path",
            "description": "CPF do cidadão (com ou sem formatação)",
            "schema": {
              "example": "123.456.789-00",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cidadão encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CidadaoResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "CPF inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Cidadão não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buscar cidadão por CPF",
        "tags": [
          "Cidadão"
        ]
      }
    },
    "/api/v1/cidadao/nis/{nis}": {
      "get": {
        "description": "Busca um cidadão pelo número do NIS (PIS/PASEP).",
        "operationId": "Cidadao_findByNis",
        "parameters": [
          {
            "name": "nis",
            "required": true,
            "in": "path",
            "description": "Número do NIS (PIS/PASEP)",
            "schema": {
              "example": "12345678901",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cidadão encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CidadaoResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "NIS inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Cidadão não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buscar cidadão por NIS",
        "tags": [
          "Cidadão"
        ]
      }
    },
    "/api/v1/cidadao/{id}/solicitacao": {
      "get": {
        "operationId": "Cidadao_findSolicitacoes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico retornado com sucesso"
          },
          "404": {
            "description": "Cidadão não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Histórico de solicitações",
        "tags": [
          "Cidadão"
        ]
      }
    },
    "/api/v1/cidadao/{id}/composicao": {
      "post": {
        "operationId": "Cidadao_addComposicaoFamiliar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateComposicaoFamiliarDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Membro adicionado com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Cidadão não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Adicionar membro à composição familiar",
        "tags": [
          "Cidadão"
        ]
      }
    },
...
    "/api/v1/cidadao/papel/cidadao/{cidadaoId}": {
      "get": {
        "operationId": "PapelCidadao_findByCidadaoId",
        "parameters": [
          {
            "name": "cidadaoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de papéis retornada com sucesso"
          },
          "404": {
            "description": "Cidadão não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar papéis de um cidadão",
        "tags": [
          "Cidadão"
        ]
      }
    },
...
  },
  "info": {
    "title": "API de Gestão de Benefícios Eventuais",
...
    "version": "1.0",
    "contact": {
      "name": "Equipe SEMTAS",
      "url": "https://www.natal.rn.gov.br/semtas",
      "email": "suporte@semtas.natal.rn.gov.br"
    },
    "license": {
      "name": "Uso Interno",
      "url": "https://www.natal.rn.gov.br/semtas/licenca"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Insira o token JWT",
        "in": "header"
      }
    },
    "schemas": {
...
      "CidadaoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "description": "ID único do cidadão"
          },
          "nome": {
            "type": "string",
            "example": "Maria da Silva",
            "description": "Nome completo do cidadão"
          },
          "cpf": {
            "type": "string",
            "example": "123.456.789-00",
            "description": "CPF do cidadão"
          },
          "rg": {
            "type": "string",
            "example": "1234567",
            "description": "RG do cidadão"
          },
          "orgaoEmissor": {
            "type": "string",
            "example": "SSP/RN",
            "description": "Órgão emissor do RG"
          },
          "dataNascimento": {
            "type": "string",
            "example": "1990-01-01",
            "description": "Data de nascimento no formato YYYY-MM-DD"
          },
          "sexo": {
            "example": "feminino",
            "description": "Sexo do cidadão",
            "allOf": [
              {
                "$ref": "#/components/schemas/Sexo"
              }
            ]
          },
          "email": {
            "type": "string",
            "example": "maria@example.com",
            "description": "Email do cidadão"
          },
          "telefone": {
            "type": "string",
            "example": "(84) 99999-9999",
            "description": "Telefone do cidadão"
          },
          "nis": {
            "type": "string",
            "example": "12345678901",
            "description": "Número do NIS (PIS/PASEP)"
          },
          "nacionalidade": {
            "type": "string",
            "example": "Brasileira",
            "description": "Nacionalidade do cidadão"
          },
          "estadoCivil": {
            "type": "string",
            "example": "Solteira",
            "description": "Estado civil do cidadão"
          },
          "escolaridade": {
            "type": "string",
            "example": "Ensino Médio Completo",
            "description": "Escolaridade do cidadão"
          },
          "empresa": {
            "type": "string",
            "example": "Empresa XYZ",
            "description": "Nome da empresa onde trabalha"
          },
          "profissao": {
            "type": "string",
            "example": "Auxiliar administrativo",
            "description": "Cargo ou função profissional"
          },
          "rendaMensal": {
            "type": "number",
            "example": 1500.5,
            "description": "Renda mensal"
          },
          "observacoes": {
            "type": "string",
            "example": "Filho com necessidades especiais",
            "description": "Informações adicionais"
          },
          "endereco": {
            "description": "Endereço do cidadão",
            "allOf": [
              {
                "$ref": "#/components/schemas/EnderecoResponseDto"
              }
            ]
          },
          "ativo": {
            "type": "boolean",
            "example": true,
            "description": "Indica se o cidadão está ativo no sistema"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "example": "2023-01-01T00:00:00.000Z",
            "description": "Data de criação do registro"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "example": "2023-01-01T00:00:00.000Z",
            "description": "Data da última atualização do registro"
          },
          "deleted_at": {
            "format": "date-time",
            "type": "string",
            "example": null,
            "description": "Data de desativação do registro, se aplicável"
          },
          "papeis": {
            "description": "Papéis que o cidadão possui no sistema",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PapelCidadaoResponseDto"
            }
          }
        },
        "required": [
          "id",
          "nome",
          "cpf",
          "rg",
          "orgaoEmissor",
          "dataNascimento",
          "sexo",
          "email",
          "telefone",
          "nis",
          "nacionalidade",
          "estadoCivil",
          "escolaridade",
          "empresa",
          "profissao",
          "rendaMensal",
          "observacoes",
          "endereco",
          "ativo",
          "created_at",
          "updated_at"
        ]
      },
      "CidadaoPaginatedResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Lista de cidadãos",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CidadaoResponseDto"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "total": {
                "type": "number",
                "example": 100
              },
              "page": {
                "type": "number",
                "example": 1
              },
              "limit": {
                "type": "number",
                "example": 10
              },
              "pages": {
                "type": "number",
                "example": 10
              },
              "hasNext": {
                "type": "boolean",
                "example": true
              },
              "hasPrev": {
                "type": "boolean",
                "example": false
              }
            },
            "description": "Metadados da paginação"
          }
        },
        "required": [
          "items",
          "meta"
        ]
      },
...
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "Código de status HTTP",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "Mensagem de erro amigável",
            "example": "Erro de validação"
          },
          "code": {
            "type": "string",
            "description": "Código de erro interno para referência",
            "example": "VALIDATION_ERROR"
          },
          "details": {
            "type": "object",
            "description": "Detalhes adicionais sobre o erro"
          },
          "errors": {
            "description": "Lista de erros de validação",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp do erro",
            "example": "2023-05-16T10:00:00.000Z"
          },
          "path": {
            "type": "string",
            "description": "Caminho da requisição que causou o erro",
            "example": "/api/v1/cidadao"
          }
        },
        "required": [
          "statusCode",
          "message",
          "timestamp",
          "path"
        ]
      },
...
      "CreateCidadaoDto": {
        "type": "object",
        "properties": {
          "papeis": {
            "description": "Papéis que o cidadão irá assumir no sistema",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PapelCidadaoCreateDto"
            }
          },
          "nome": {
            "type": "string",
            "example": "Maria da Silva",
            "description": "Nome completo do cidadão"
          },
          "cpf": {
            "type": "string",
            "example": "123.456.789-00",
            "description": "CPF do cidadão"
          },
          "rg": {
            "type": "string",
            "example": "1234567",
            "description": "RG do cidadão"
          },
          "data_nascimento": {
            "format": "date-time",
            "type": "string",
            "example": "1985-10-15",
            "description": "Data de nascimento do cidadão"
          },
          "sexo": {
            "type": "string",
            "enum": [
              "masculino",
              "feminino",
              "outro"
            ],
            "example": "feminino",
            "description": "Sexo do cidadão"
          },
          "nis": {
            "type": "string",
            "example": "12345678901",
            "description": "Número de Identificação Social (NIS) do cidadão, utilizado para programas sociais. Obrigatório para beneficiários."
          },
          "nome_social": {
            "type": "string",
            "example": "Maria Santos",
            "description": "Nome social do cidadão (usado para pessoas trans e travestis)"
          },
          "telefone": {
            "type": "string",
            "example": "(84) 98765-4321",
            "description": "Telefone do cidadão para contato"
          },
          "email": {
            "type": "string",
            "example": "email@exemplo.com",
            "description": "Endereço de email do cidadão para contato"
          },
          "endereco": {
            "description": "Endereço do cidadão",
            "allOf": [
              {
                "$ref": "#/components/schemas/EnderecoDto"
              }
            ]
          },
          "renda": {
            "type": "number",
            "example": 1200.5,
            "description": "Renda mensal do cidadão em reais (R$)"
          }
        },
        "required": [
          "nome",
          "cpf",
          "rg",
          "data_nascimento",
          "sexo",
          "endereco"
        ]
      },
      "UpdateCidadaoDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Maria da Silva",
            "description": "Nome completo do cidadão"
          },
          "rg": {
            "type": "string",
            "example": "1234567",
            "description": "RG do cidadão"
          },
          "data_nascimento": {
            "format": "date-time",
            "type": "string",
            "example": "1985-10-15",
            "description": "Data de nascimento do cidadão"
          },
          "sexo": {
            "type": "string",
            "enum": [
              "masculino",
              "feminino",
              "outro"
            ],
            "example": "feminino",
            "description": "Sexo do cidadão"
          },
          "nis": {
            "type": "string",
            "example": "12345678901",
            "description": "Número de Identificação Social (NIS)"
          },
          "telefone": {
            "type": "string",
            "example": "(84) 98765-4321",
            "description": "Telefone do cidadão"
          },
          "email": {
            "type": "string",
            "example": "maria.silva@email.com",
            "description": "Email do cidadão"
          },
          "endereco": {
            "description": "Endereço do cidadão",
            "allOf": [
              {
                "$ref": "#/components/schemas/EnderecoDto"
              }
            ]
          },
          "renda": {
            "type": "number",
            "example": 1200.5,
            "description": "Renda mensal do cidadão"
          }
        }
      },
...
    }
  }
}